Wrong tense
Some of the validation errors are actually in the future tense (weird). "deverá" changed to "deve".
Also small fixes on some sentences.
And fixes on the use of the word "conter" and "ter" the first is a property inside something the second is a property of something.
Also changed the array validation to return "array" instead of "conjunto" it's better since we don't use this word often in brazil, also the string validation is currently returning "string" instead of "cadeia de caracteres" so now both have the same aproach.